refactor(anchor): consts and utils
