+++ packages/components/anchor/anchor-item.tsx
@@ -2,7 +2,7 @@ import { defineComponent, h, VNodeChild, onMounted, onUnmounted, inject, watch }
 import { ANCHOR_SHARP_REGEXP } from './utils';
 import props from './anchor-item-props';
 import { usePrefixClass, useCommonClassName } from '../hooks/useConfig';
-import { AnchorInjectionKey } from './constants';
+import { AnchorInjectionKey } from './consts';
 import { isFunction } from 'lodash-es';
 import { isString } from 'lodash-es';
