+++ packages/components/anchor/anchor.tsx
@@ -17,7 +17,7 @@ import { SlotReturnValue } from '../common';
 import Affix from '../affix';
 import { TdAnchorProps } from './type';
 import { usePrefixClass, useCommonClassName } from '../hooks/useConfig';
-import { AnchorInjectionKey } from './constants';
+import { AnchorInjectionKey } from './consts';
 
 export interface Anchor extends ComponentPublicInstance {
   scrollContainer: ANCHOR_CONTAINER;
